fix(candid): count review items in a type that cannot wrap, and read vec lengths as LEB128
Two defects that meet in the same range of values.

The number of review items is a fixed base plus a repeat count taken from the
payload, computed in a uint8_t. A large enough vec rolled the total over: with
255 followees, 3 + 255 became 2, so the device drew two items for a transaction
carrying 255 elements and signed the rest without showing them. The counters are
now uint16_t and parser_getNumItems refuses a total it cannot address rather
than truncating it. Counts that wrapped to zero already failed closed on the
existing "> 0" assertion; this closes the ones that wrapped to a small non-zero
number.

The follow case in the report cannot reach the chain -- NNS governance caps
followees at 15, so an oversized Follow is rejected after signing. The SNS
AddNeuronPermissions path is the one that matters: a list size of 253 to 255
wraps the count to 1 to 3, which stops the review before the item carrying the
grantee Principal Id, on a canister that is not pinned. The device would show
the type, canister and neuron while signing a permission grant to a principal
the user never saw.

Separately, a vec's element count is LEB128, so any count from 128 up occupies
more than one byte, and readCandidByte consumed only the first. That mis-read
the count and left the parser one byte behind the encoder, so the elements were
walked from the wrong offset while the replica read the same payload correctly
-- a display-versus-signed divergence in exactly the range that also overflowed
the arithmetic above. readCandidVecLength decodes the whole length and rejects
one that will not fit the uint8_t field it is stored in.

Deliberately not capping repeat counts at a fixed number. NNS caps followees at
15 but the SNS permission list has no such published bound, so a hard cap risks
refusing payloads the chain accepts. Refusing only what cannot be displayed
keeps the device from rejecting valid work.

Tests: LEB128 lengths of 0, 1, 127, 128, 200 and 255 decode to the right value
and leave the offset past the whole length, checked with a sentinel byte so a
short read is caught; lengths of 256 and above are refused; a truncated length
errors. All three fail against the previous single-byte read. The existing 1041
tests are unchanged.

+++ app/src/parser_impl.h
@@ -38,7 +38,7 @@ parser_error_t _readTransactionStateRead(const parser_context_t *c, parser_tx_t 
 
 parser_error_t _validateTx(const parser_context_t *c, const parser_tx_t *v);
 
-uint8_t _getNumItems(const parser_context_t *c, const parser_tx_t *v);
+uint16_t _getNumItems(const parser_context_t *c, const parser_tx_t *v);
 
 parser_error_t getManageNeuronType(const parser_tx_t *v, manageNeuron_e *mn_type);
